Adjust path for sourcing .env file
The code has been updated to use an absolute path for the .env file, instead of a relative one. In all scripts, the changes ensure that .env is found and sourced correctly, regardless of from where the scripts are run. This adds flexibility and reliability to the execution of scripts 'mysql', 'mysqldump', and 'de'.